Split test_marvin_actions into multiple files
One file per action, in preparation of decoupling the actions and make
them plugins.
